Refactor services to reduce number of classes
