+++ LightBulb/LightBulb.csproj
@@ -77,15 +77,11 @@
     <Compile Include="Models\GeoLocation.cs" />
     <Compile Include="Models\HotKey.cs" />
     <Compile Include="Models\CycleState.cs" />
-    <Compile Include="Models\SolarInfo.cs" />
-    <Compile Include="Services\GammaService.cs" />
-    <Compile Include="Services\ColorTemperatureService.cs" />
-    <Compile Include="Services\GeoLocationService.cs" />
-    <Compile Include="Services\HotKeyService.cs" />
+    <Compile Include="Services\CalculationService.cs" />
+    <Compile Include="Services\LocationService.cs" />
     <Compile Include="Services\SettingsService.cs" />
-    <Compile Include="Services\SolarInfoService.cs" />
     <Compile Include="Services\UpdateService.cs" />
-    <Compile Include="Services\WindowService.cs" />
+    <Compile Include="Services\WinApiService.cs" />
     <Compile Include="ViewModels\Components\AdvancedSettingsViewModel.cs" />
     <Compile Include="ViewModels\Components\GeneralSettingsViewModel.cs" />
     <Compile Include="ViewModels\Components\HotKeyViewModel.cs" />
